a2a-with-agentcore/agents/orchestrator/requirements.txt
strands-agents>=1.4.0
bedrock-agentcore>=0.1.2
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
pydantic>=2.0.0
sse-starlette>=2.0.0
aws-opentelemetry-distro>=0.10.0
